pyproject.toml
[tool.poetry]
name = "repl-nix-journeylogger"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
flask-sqlalchemy = "^3.1.1"
flask-login = "^0.6.3"
werkzeug = "^3.0.4"
psycopg2-binary = "^2.9.9"
weasyprint = "^62.3"
pytz = "^2024.2"
flask-wtf = "^1.2.1"
python-magic = "^0.4.27"
reportlab = "^4.2.5"
wtforms = "^3.1.2"
openai = "^1.51.0"
flask-migrate = "^4.0.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
